refactor(node): Move Yarn2 into its own dedicated package
Prepare for adding further Yarn2 specific classes / files.
